test: Make CLI tests to integration tests instead of unit tests
